Apache scanner v3
Added documentation, moved to the services subdirectory and fixed
some minor issues.
